refactor: Follow TODOs to convert from template
The template this project was created from contains several TODOs that
need following to create the service correctly. Follow and remove the
TODOs.

TISNEW-4304
