+++ README.md
@@ -1,33 +1,17 @@
 # TIS Microservice Template
 
 ## About
-This is a template to be used for TIS microservices with the following
-technology:
-
- - Java 11
- - Spring Boot
- - Gradle
- - JUnit 5
-
-Boilerplate code is to be generated with:
- - Lombok
- - MapStruct
-
-Code quality checking and enforcement is done with the following tools:
- - EditorConfig
- - Checkstyle
- - JaCoCo
- - SonarQube
-
-Error and exception logging is done using Sentry.
+This service handles synchronization of data from TIS Core to TIS Self Service.
 
 ## Usage
-To use this template, create a new repository from it and follow the TODOs in
-the code, with the following additional changes.
- - Update copyright year in [LICENSE](LICENSE).
- - Update copyright year in [TemplateApplication].
- - Update copyright year in [TemplateApplicationTest].
- - Update this README.
+The service can be started using the `bootRun` Gradle task with the included
+Gradle wrapper `./gradlew bootRun`.
+
+Spring Actuator is included to provide health check  and info endpoints, which
+can be accessed at `<host>:<port>/sync/actuator/health` and
+`<host>:<port>/sync/actuator/info` respectively.
+
+## Deployment
  - Provide `SENTRY_DSN` and `SENTRY_ENVIRONMENT` as environmental variables
    during deployment.
 
@@ -36,6 +20,3 @@ This project uses [Semantic Versioning](semver.org).
 
 ## License
 This project is license under [The MIT License (MIT)](LICENSE).
-
-[TemplateApplication]: src/main/java/uk/nhs/hee/tis/template/TemplateApplication.java
-[TemplateApplicationTest]: src/test/java/uk/nhs/hee/tis/template/TemplateApplicationTest.java
