+++ src/main/resources/application.yml
@@ -1,5 +1,4 @@
 server:
-  # TODO: Update server port and context path to a unique value.
-  port: 8080
+  port: 8208
   servlet:
-    context-path: /template
+    context-path: /sync
